[compiler] Set attributes on packed args and on loads from them
This commit lets LLVM know that the pointer to the packed argument
structure may not be null, must not be undef/poison, and is
dereferenceable.

It also transfers `noundef` and `nonnull` attributes from the old
parameters to the new loads from the argument struct. Those loads can
take `!noundef` and `!nonnull` metadata.

This should improve performance in certain cases, as this pass typically
runs before the final O3 optimization pipeline and any extra information
we can give LLVM should help.