Bump GoogleTest.
The old version of GoogleTest we were using uses macros in a
non-C++17-conforming way. This has been improved (though not fully
fixed) in more recent versions.

Unfortunately, we were using an internal GoogleTest function that we
should have left alone. This function still exists, but is now static
and uses a new type in an anonymous namespace, so cannot be referred to
from external code without patching GoogleTest. We have locally patched
GoogleTest before when it did not quite work for us, so this patches it
again.