Support LLVM already being set up in OCK rather than importing it.
OCK previously required CA_LLVM_INSTALL_DIR to be set in order to
import LLVM. If OCK is included from LLVM, then we don't need to do
this. We can do this by checking if LLVM CMake variables are set first.

We also change to use the more general LLVM method of using
add_include_dirs rather than adding target include for LLVM, as that
has issues with using OCK through FetchContent (or as an external
LLVM project).