[compiler] Provide a safe GDB JIT registration listener
This commit adds a version of LLVM's GDBRegistrationListener that is
safe to use in our libraries.

This and the original listener is responsible for registering and
deregistering dynamically compiled kernels with GDB's JIT compilation
interface. It does this through access to global symbols understood by
GDB. This requires that it must be made correctly thread-safe.

LLVM ensures this by having only one static singleton version of its GDB
listener, which locks itself when notified of a new object, or notified
or a freed one. This does not work for our use-case, as the static
singleton may be destroyed before the OpenCL library is destroyed. Thus
our modules and kernels may inadvertently use an already freed object.

To solve this, our version of this listener is no longer a singleton,
and instead requires the callers to correctly lock access to it. In
principle this requires the host target to lock when jitting a kernel
and when freeing a kernel. Both of these are now locked under the same
LLVM global mutex to ensure thread safety.

The original also accesses these globals in its destructor, but our
version requires (asserts) that resources are already freed up at the
time the listener is freed. This requires us to lock in one fewer
instance and should simplify the mental model of how the listener works.