[UnitCL] Test *Half* tests under the 'COMPILER' filter
These fp16 tests are important enough that we should be running them in
more configurations more often - with vectorization, with different
compilation flags, etc. Relegating them to only the `check-UnitCL-half`
target means we miss out on this coverage.

These will all be skipped if half/fp16 isn't enabled so shouldn't
negatively impact testing time in the majority of configs.