Fix in-tree builds' use of LLVM tools for compiling builtins
This looks like it was intended to work, by checking whether or not the
`$tools_dir` was empty. Unfortunately, unless the tools directory is
manually set to empty by a user, it will always be at least `/bin`. This
meant that in practice is was trying to pick up tools from the path,
leading to unintended behaviour.

This commit fixes this behaviour, by only setting
`CA_BUILTINS_TOOLS_DIR` if `CA_LLVM_INSTALL_DIR` is set.

If in-tree targets are used, we set the CMake targets using an alias
executable. This gives us the full path to the in-tree target executable
without worrying about accidentally using `clang` from the PATH, for
example. This fixes another latent bug in the old code, which would
import a `builtins::compiler` target called "clang" which may actually
be on the PATH.

It also:

* deletes dead code:
  * FindLLVMTool was never used
  * the regex checking of LLVM tool binary names was always trivially
    matching because we set the variables ourselves
  * we never used the `BUILTINS_*` variables we exported from the macro
* consolidates the setting and use of `CA_BUILTINS_TOOLS_DIR` in one
  place
* removes references to `opt` which we no longer use when compiling
  builtins.