[spir-v] Bump SPIRV-Headers
A recent change to the upstream SPIRV-Headers made a non
backwards-compatible change to the name of some tokens. This commit
bumps us to just past that change.

This becomes a problem when building OCK as part of an in-tree DPC++
build, because DPC++ makes an earlier call to FetchContent_Declare on
the same SPIRV-Headers project, and that version is cached, meaning we
pick up on that version of the repo - one we haven't asked for.

While this isn't ideal that we are not in control of the headers,
there's no reason for us to be languishing on an older version of
SPIRV-Headers when newer ones should always be 'more correct'.

Unfortunately they haven't made any tags since August last year and we
need something newer than that, so I've arbitrarily chosen the most
recent commit as of the time of writing.