Add RISC-V to the host target supported platforms
This gives support for the Host target's clc to support RISC-V as an
additional compiler as well as building OCK host target for RISC-V as a
cross compile.

Note this also disables the JIT path as RISC-V does not support this.
We no longer go through the JIT constructs to create the TargetMachine
if deferred compilation is disabled.