Fix check-ComputeAorta depending on check-cross-ComputeAorta
A recent refactor accidentally made the global `check` target depend on
`check-cross-ComputeAorta` which, because `check-ComputeAorta` depends
on `check`, running `check-ComputeAorta` would run
`check-cross-ComputeAorta`.

This in theory shouldn't be a problem but the `check-cross-ComputeAorta`
target looks to make incorrect assumptions about which test/check
targets are available, and unconditionally depends on them even when
they may not exist.

This is problematic and so we should look at fixing the 'cross' target,
but in the meantime we need to reinstate the correct intended dependency
chain on the two check targets.