Unify check target names
This commit attempts to streamline and make consistent our check target
names, so that they are unconditionally named `check-ock-*` in all build
configurations.

The generic `check` target still exists, but now it is an alias of
the new global `check-ock` target, which depends on all the internal
check targets such as `check-ock-UnitCL` etc.

This makes it easier to splice out the `check` target, and have
everything remain internally consistent. It also means users and
developers can consistently run `check-ock-*` targets without worrying
about which build configuration they are in.

Notable exceptions are the now legacy `check-ComputeAorta`-style
targets, which now alias (possibly new) `check-ock` versions:

* `check-ComputeAorta` aliases `check-ock`
* `check-tidy-ComputeAorta` aliases `check-ock-tidy`
* `check-cross-ComputeAorta` aliases `check-ock-cross`

We couldn't afford to rename `check-ComputeAorta` as
`check-ock-ComputeAorta` using the same internal helper functions, so
these old targets are now handled specially via custom targets.

This should help us slowly move to the new targets and eventually remove
the 'ComputeAorta' styling, as it is no longer the official name of the
OCK.

I've also made it so the `check-ock-tidy` target now requires the `tidy`
target to be present and the `check-ock-cross` target now also requires
all of its dependencies to exist, which should fix build issues in
certain configurations.

All documentation which referred to `check-` or `check-ock` can now be
unified to assume `check-ock`. I've still left in the internal lack of
assumptions about the precise name, since we have the infrastructure for
it and it should make the build system less brittle, but in documentation
it is probably better to spell out explicitly, on balance.