[cl] Support multi-device testing for CL examples
This commit fixes the testing of CL examples in a multi-device OCK
build. Previously each example was provided only a platform and so the
example was asking for a device over standard input, which we couldn't
provide and thus the test failed.

Now each example is given a check target for each device enabled, where
the device is provided to the example.

Each of these device-specific check targets is added to a group for each
example, meaning all devices can easily be run for each device. It also
preserves the old behaviour for a single-device build, wherein the same
check target runs the example with the (only) enabled device.

The example's group is added to the global CL-specific check, as well as
the global check target. This means each example is only run once when
running one of the global check targets.

For example:

* both `check-ock` and `check-ock-cl` run:
  * `check-ock-clVectorAddition` which runs:
    * `check-ock-clVectorAddition-ComputeAorta_x86_64`
    * `check-ock-clVectorAddition-RefSi_G1_RV64`