Fix building of CL examples when CA_ENABLE_TESTS=OFF
A previous change made the mistake of thinking that the examples weren't
built at all when `CA_ENABLE_TESTS` was off, when in fact the examples are
unconditionally built, but when `CA_ENABLE_TESTS` is off, the check
targets are not built.

The trick is that `add_ca_cl_check` does nothing when `CA_ENABLE_TESTS`
is off, so the targets that the group names depend on are not present.