Fix signed integer overflow in add_sat.
This fixes an error seen in UBSAN builds. If x < (T)0, then min - x
cannot overflow, but if x >= (T)0, it can. We were relying on the result
of that overflow being ignored by masking it out, but it is undefined
behavior to even perform the operation, so we need to avoid that.
Unsigned subtraction is well-defined for any values; we can use that
instead.