[oclc] Fix input and output of 'half' values
Internally we were storing half data as cl_half, which is correct, but
this is not an actual floating-point data type. Therefore when we
converted those values to/from floating-point types, things would go
wrong (e.g., we'd static cast 2.0 to the integer 2 instead of 0x4000).

The OpenCL headers provide some methods for correct translation between
float and half, so we can use those instead.