[lld] Fix LLD library dependencies
The dependencies on the LLD libraries were not being resolved by CMake
because we were linking against them not by their target names but by
a string denoting their fully qualified path. This meant that CMake
couldn't order the static libraries according to the dependency chains
and it was random chance whether we managed to link successfully.

This issue was made manifest when building the host (with RISC-V
enabled) and riscv targets together. The host compiler - being processed
first - added a dependency on libLLVMRISCVCodeGen.a. The riscv target
coming second then added a dependency on compiler-linker-utils which in
turn had a dependency on libLLVMRISCVCodeGen.a.

Since a dependency on that library was already registered, it (and its
dependencies like libLLVMCodeGen.a) came first in the link order before
liblldCommon.a, which has a dependency on the same library. Manually
ordering libLLVMCodeGen.a in compiler-linker-utils didn't work because
CMake would prune the "duplicated" dependency.

By correctly linking using the libraries' target names, CMake can detect
the dependencies and order libLLVMCodeGen.a after both of the libraries
that depend on it.

Also while debugging the issue I noticed that the `compiler-riscv`
library was linking against `compiler-riscv-utils` and a bunch of other
libraries which `compiler-riscv-utils` itself was linking against. Since
`compiler-riscv` doesn't use/export any symbols of its own, and it
publicly links against `compiler-riscv-utils`, these dependencies are
unnecessary.