Fix aarch64 failure due to flags for non-jit path
The cross aarch64 "host" target was crashing when running the
offline test
Execution/Execution.Barrier_02_Barrier_No_Duplicates/OfflineOpenCLC.

This was found to be due to the flags passed to the creation of the
TargetMachine being different to what is passed from the
JITTargetMachineBuilder. The main ones that seemed to matter was
the CodeModel and the `JIT` parameter.

This change uses the defaults for the CodeModel and RelocationModel
and also sets the JIT parameter. The need for the JIT parameter should
be investigated to understand why the failure happens.