[host] Add initial support for RISC-V relocations
This adds support for RISC-V relocations - enough to permit the running
of UnitCL (in the default configuration). There are obviously many other
relocations for RISC-V but without test coverage those have been left
for future work.

Note that relocation support is still incomplete. There is no support
yet for the PLT or GOT sections, so if the executable is loaded too far
away from external symbols, the relocation process will fail at program
load time, complaining of an unreachable symbol. This situation is
entirely down to the in-process memory of the OpenCL location and the
machine it's running on, so cannot be easily pre-determined.

Locally, at least, all (offline) UnitCL tests pass in isolation but fail
when run together, because the binary is loaded too far away from some
symbols such as memcpy.

The ADD32 and SUB32 relocations only seem to appear for the PIC
relocation model, which is not being used in main nor in this PR, but
they are simple enough that including them should be fairly obviously
correct.