Ignore test_gl test.
This test covers the cl_khr_gl_sharing extension, but is not skipped on
implementations that do not support this extension.

At the moment we comment out this test in our internal OpenCL CTS
testing, rather than using the ignore file for it, but this causes
problems when we move to using OpenCL CTS's own CSV files.