Fix vkEnumerateInstanceExtensionProperties.
GCC 13 rightly complains about an out of bounds pointer: when
*pPropertyCount > vk::instance_extensions.size(), merely computing
vk::instance_extensions.begin() + *pPropertyCount has undefined behavior
and trying to cover that up via std::min afterwards is too late and not
guaranteed to do what is intended.

While looking into this I also found that we did not match the spec in
that we did not return VK_INCOMPLETE when the passed in size was too
small, and did not set *pPropertyCount when the passed in size was too
large.