Fix LLVM assertion about wrong debug info
When an inlineable call is emitted in a function that has debug info,
that call must itself have debug info. In AddFloatingPointControl, in
configX86FP, we made sure to call a utility function that sets up the
debug info, but in configAarch64FP and configArmFP we did not do the
same, resulting in errors.