Remove DisableNeonAttribute pass.
We had a DisableNeonAttribute pass which applied the -neon attribute to
work around a specific issue in Neon codegen in LLVM 4, that we do not
know to still be an issue. Meanwhile, we do know that applying -neon
causes issues: we get errors during instruction selection when Neon
intrinsics are used while Neon is not supported.