[UnitCL] Fix C11Atomics_76_*, C11Atomics_77_*, C11Atomics_78_*
These tests test atomic_compare_exchange_weak_explicit which is allowed
to fail for no clear reason. The test tries to account for this by
detecting output differences, using that to fill a weak_exchanges_failed
vector, and using that weak_exchanges_failed vector to detect where the
result of atomic_compare_exchange_weak_explicit does not match
expectations. That is, in output_reference, it does:

          if (value == expected_values[success_indices[index]]) {
            weak_exchanges_failed[index / kts::localN] = true;
            return true;
          }

Then, in bool_output_reference, it does:

          const size_t work_group = index / kts::localN;
          const size_t success_index_of_workgroup = success_indices[work_group];
          return (index == success_index_of_workgroup) &&
                 !weak_exchanges_failed[work_group];

The problem here is the different meanings of index. In
bool_output_reference, index is the global ID, and the division by
kts::localN is to convert it to a workgroup ID. But in output_reference,
index is already a workgroup ID, and the division by kts::localN causes
the wrong element of weak_exchanges_failed to be set.

This commit renames index to either wgid or gid, to make it clearer
which index is being used, and takes out the incorrect division by
kts::localN.