Specify dependency on Abacus headers.
The previous fix was insufficient to avoid clang-tidy running before
Abacus headers were generated.

This commit adds a DEPENDS keyword to add_ca_tidy, add_ca_library, and
add_ca_executable. If specified, the dependencies are added to the main
target, the tidy target, and the tidy target's symbolic files. Existing
dependencies on abacus_generate are updated to use DEPENDS so that the
clang-tidy commands will pick this up.