Report availability of out of order exec mode.
If CA_ENABLE_OUT_OF_ORDER_EXEC_MODE is enabled (as it is by default), we
accept CL_QUEUE_OUT_OF_ORDER_EXEC_MODE_ENABLE in queue properties, but
do not report it as a property we support for queues. This commit
rectifies that.

This is not currently tested by OpenCL CTS, but will soon be.