Set LLVM_DIR.
When we import LLVM, we do so by including LLVMConfig directly, which
works fine. However, we do the same for LLD, we include LLDConfig, but
LLDConfig needs LLVM and does find_package(LLVM 19.0.0 [...]). Although
it sets a hint path here that would pick up the previously included
LLVM, the hint path is ignored in cross builds with
CMAKE_FIND_ROOT_PATH_MODE_PACKAGE set to ONLY, which it normally should
be set to because host packages cannot be used for the target. By
setting LLVM_DIR, we allow the cross LLVM to be found in the expected
location, outside of CMAKE_FIND_ROOT_PATH.