Do not skip events.
In clEnqueueBarrierWithWaitList, clEnqueueMarkerWithWaitList, and
clEnqueueWaitForEvents we were assuming that all events to wait for
would be from the same queue, and therefore implicit dependencies would
take care of it already. There is no such requirement, events to wait
for may be for different queues, they are merely required to have the
same OpenCL context.

No UnitCL test is added as this is difficult to test for reliably, and
was caught by the OpenCL CTS already.