Do not mark events as running.
We were previously marking events as running before they were running.
If we have commands A, B, with implicit dependencies, we would mark both
of A, B as running once they were submitted. As OpenCL CTS test_events
reveals, this is incorrect, because it allows an application to retrieve
the execution status of B, determine that it is running, and then
retrieve the execution status of A, and determine that it is running as
well, when it is only possible for B to start running once A is
completed. Never marking commands as running is the minimal change to be
a poor-quality conforming implementation (commands are never guaranteed
to take long enough for the "running" state to be observed). This is
intended to be improved in a future commit so that we do mark events as
running again, but only when they actually are running.