Minor changes to cpu hal to enable wrapping for remote hal access
Wrapping of the cpu hal with a server and remote access from a client
requires some changes to make it work. These are:

The creation of the device info being a function that can be called
externally through a static function. This allows us to use clc without
having to use the client/server relation to get the device info.

Making the linker script (which is empty) defined within the static function
rather than baking it. Technically we could bake it but we wanted to
simplify the server and build the cpu hal files directly rather than
having to build it as a shared library and complicate the dependencies.

Also added clik as a reason to run pr testing as the clang format needs
to be run.