[CMake] Fix check for LLVM_ENABLE_ZLIB, add _ZSTD.
In cross builds, we check for a LLVM_ENABLE_ZLIB mismatch between host
LLVM and target LLVM because it affects bitcode compatibility. This
mismatch was not fully reliable: LLVM sometimes sets LLVM_ENABLE_ZLIB to
other values than the values we check for, in particular it may set it
to nothing. This change allows any arbitrary value to be picked up.

In cross builds, LLVM_ENABLE_ZSTD has the same problem as
LLVM_ENABLE_ZLIB: if the host LLVM has that enabled, the target LLVM
needs it too. So do the same check for that.