[RISC-V] Link FP64 and FP16 options to d and Zfh.
Previously, our list of enabled extensions was hardcoded and when a user
would set CA_HOST_ENABLE_FP16 on RISC-V, it would fail because it would
use softfloat and attempt to use compiler-rt/libgcc functions that we do
not provide. But presumably, when CA_HOST_ENABLE_FP16 is enabled, the
user wants to use Zfh half-precision hardware instructions, so enable
the extension instead. And do the same for FP64, where the extension was
previously always enabled, but now gets linked to CA_HOST_ENABLE_FP64.