[host] Remove invalid caching of object code.
As seen in clBuildProgramTwiceTest.RedefineMacro, we may reuse a
HostModule to create a binary from a different LLVM module, or with
different LLVM build options. We do not have sufficient information to
determine whether we can re-use a prior result, and can reasonably
assume that if createBinary is called a second time, it will be called a
second time specifically because the caller does not want the prior
result.