Fix memory leak.
In bf1b1eab5b, I cleared host->ndranges, but the result of that the loop
in hostDestroyCommandBuffer to free the arguments no longer does
anything useful: there is no longer any reference to the arguments.

Change packed_args to a mux::dynamic_array so that this is automatically
taken care of for us.