Run AlignModuleStructsPass earlier.
The AlignModuleStructsPass patches up structs according to OpenCL
alignment requirements so that they behave the same way on all
implementations. We were already calling this, but we were calling it
after running a number of upstream LLVM passes, so we were feeding them
incorrect input and hoping we could correct that afterwards. A change
has gone in on LLVM 19 that performs a valid optimization that we cannot
correct afterwards, so make sure we run AlignModuleStructsPass as soon
as we know what target we are compiling for.