Fix native builtins with double as the input.
If double was used as the input for the native builtins then we would
get a link error. native builtins don't have an overload for double
but we were providing one. This was also true of some half native builtins.

Removing from the overloads fixes the issue as double or half are then
cast to float.