Fix printf of nan.
std::string::replace takes pos and count parameters, not start and end.
This was not previously covered by testing because either the format
specifier appeared at the start (meaning count and end are the same), or
the format specifier appeared at the end (meaning there are no
characters after the format specifier to cut off). Add a test that puts
the format specifier in the middle.