Fix risc-v memmove creating memcpy instrinsic
SYCL CTS test vector_load_store with host risc-v showed that memmove
intrinsics could result in a linked builtin which then included memcpy
instrinsics which could result in a linker error.

This resolves this by passing -fno-builtin in the build for the builtins.
This only removes the additional llvm.memcpy and all other instrinsics
remain.