Fix risc-v memmove creating memcpy instrinsic
SYCL CTS test vector_load_store with host risc-v showed that memmove
intrinsics could result in a linked builtin which then included memcpy
instrinsics which could result in a linker error.

This resolves this by passing the attribute
__attribute__((always_inline, no_builtin)) to the memmove builtin's call
of memcpy and adds lit tests for memcpy,memmove and memset intrinsics.