[MSVC] Fix debug builds.
We build Abacus in debug builds with /RTCc to flag out of range
conversions, but this is a leftover from when Abacus was its own
project. In oneAPI Construction Kit, it does not make sense and does not
work, as the infrastructure to report such conversions is not available.
It was also completely untested as it was only enabled in Debug builds,
which we never ran.