Fix UnitCL file descriptor leak.
For testing printf, we were saving the original stdout in a new fd,
redirecting stdout to a temporary file, running the test, restoring stdout
from that fd, and reading from that file. What is missing here is closing
that new fd after we are done with it, so we would open more and more and
more file descriptors. On macOS, the file descriptor limit is lower than it
is on Linux and we would run out of file descriptors.