Auto-disable deferred compilation.
CA_HOST_DUMP_ASM=1 and deferred compilation are incompatible. Our
current behavior is that CA_HOST_DUMP_ASM is effectively silently
ignored, making for a poor user experience.

With this change, when CA_HOST_DUMP_ASM=1 is set and deferred
compilation would normally be implicitly enabled, deferred compilation
will be disabled and a warning is printed for it.

If an explicit CA_HOST_DEFERRED_COMPILATION=1 is set, deferred
compilation can be forced. This still works and results in no assembly
being printed.