[UnitCL] Fix C11Atomics_67*...C11Atomics_72*
As with the later C11Atomics_* tests, these tests try to account for
atomic_compare_exchange_weak sporadically failing, as it is permitted
to, by detecting output differences and using that to determine whether
atomic_compare_exchange_weak's result must have been true, or must have
been false. However, unlike the later C11Atomics_* tests, these tests
were not ensuring that the random data was unique. Therefore, it was
possible to wrongly detect that an atomic_compare_exchange_weak
operation must have succeeded because the data was equal to the data
that it tried to set it to, except the operation had actually failed
because the old data was, by random chance, already equal to the data
that it tried to set it to. Because of how our random integer number
generation works, there would be at least one element already set to its
desired value, across all affected tests, in almost every UnitCL run.

To fix this, ensure the values are unique just like later tests do.