[RISC-V] Make mem* functions always available
When we link in builtin functions, we attempted to detect when we need
memcpy/memmove/memset to provide them when needed. This detection proved
to not be entirely reliable: compiler passes run after linking in the
builtin functions are permitted to insert additional references. When
running SYCL-CTS, it turned out that function inlining can insert
additional references to memcpy for the purpose of copying function
arguments which, after inlining, no longer exist as separate function
arguments.

For non-RISC-V targets, we already unconditionally provide these
functions in host::utils::getRelocations(), so it seems reasonable to
just always provide them for RISC-V too.

This also allows us to avoid some complexity in the definition of
memmove that was only there to deal with the situation where memcpy
might be unavailable. As that can no longer happen, that no longer needs
to be taken into account.