lldLinkToBinary: disable subreg liveness.
On LLVM versions prior to LLVM 20, enabling subreg liveness would lead
to invalid code generation if an instruction defined an early-clobber
register, used another register, and the register allocator wrongly
allocated the subregister of the other register as the early-clobber
def. Work around this by disabling subreg liveness.

This is done in lldLinkToBinary so that it does not affect host: at the
moment, we only know this to be a problem with RISC-V, and only when RVV
is enabled. On host, we do not yet provide a supported way of enabling
RVV.