[NFC] Rename Abacus add_exact functions.
Elsewhere in Abacus, we use f_unsafe as a convention for a version of f
that makes specific assumptions about its input parameters, and f for
the version that does not make those assumptions. For add_exact, we had
it the other way around, we had add_exact make specific assumptions
about its input parameters, and add_exact_safe which did not.

Rename add_exact to add_exact_unsafe, rename add_exact_safe to
add_exact. For consistency, provide the same overloads for both.