Add ock artefact building, icd building and tornado run for planned testing
Added planned testing which is intended to run long running scheduled
tests. Added a base which they can all use which allows all jobs to
follow a similar method of using a matrix, a named target and a flag
which dictates what tests are enabled.

icd, ock artefacts and tornado are done as part of this.

Fixed github action ninja to be a set hash