[AlignModuleStructsPass] Update types in constants
Execution.Regression_35_Constant_Struct_Alignment reveals on LLVM 20
that if a constant expression references a type that needs to be
replaced, we were ignoring it unless the constant expression also
happened to contain references to globals that need to be replaced. Do
a pass over constant operands and update them to instructions as needed.