Improve x86-linux build support.
We provide CA_BUILD_32_BITS to attempt to build a 32-bit version of
oneAPI Construction Kit, but this is fragile and does not always work.
Notably, this has stopped working in our internal CI because CMake
sometimes fails to pick up libraries required by LLVM.

Regular cross compilation which we use for other targets is more
reliable. To allow this, add an x86-toolchain.cmake file.

This also reveals that our x86 detection was incomplete and would result
in oneAPI Construction Kit failing to build on native 32-bit platforms,
because CMAKE_SYSTEM_PROCESSOR will not be "x86", it will be e.g.
"i686". Account for this in the detection.