Builder::create<OpVariable>: Set address space of builtins.
Builtins are supposed to be declared with the Input storage class, but
are not. We were working around this by ignoring the storage class and
forcibly overriding how the builtin got declared, but this breaks now
that DPC++ generates different SPIR-V, which puts it in a different
storage class, resulting in us seeing address space casts where there
should be none. Work around this by putting globals in whatever address
space is implied by the storage class.