Fix logic to query the number of available devices
For the riscv mux target, GetDeviceInfos always returned 1 when called
with device_infos_length = 0 and out_device_infos = nullptr to query
the number of available devices, regardless of the actual number of
devices detected in the system.

This fixes it by always returning the number of available devices in the
system, but still only filling the provided vector up to its own
capacity.