Ensure that expected fails on CTS always show as a pass.
When we have regular fails which we do not expect an immediate fix, we
want to show as pass so we can catch be failures more easily.

When we know there is an expected fail, we create files to cover fails
on all targets or fails on a particular target for OpenCL CTS and
SYCL CTS. These will be under .github/opencl_cts and .github/sycl_cts.
Fails files will be of the form expect_fail_all.csv or expect_fail_<target>.csv.
These should match lines in the original csv file being used.